feat: Replace URLs with Link in session details
